requirements.txt
aiogram>=3.22.0
flask==3.0.3
flask-login==0.6.3
flask-caching==2.1.0
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
werkzeug==3.0.3
python-dotenv==1.0.1
aiosqlite==0.20.0
cachetools==5.3.3
pillow==10.4.0
openpyxl==3.1.5
cryptography==42.0.8
ujson==5.10.0
asyncio-throttle==1.0.2
prometheus-client==0.20.0
faker==26.0.0
aiohttp>=3.12.0
psutil==6.0.0
aiohttp-wsgi==0.8.2
requests==2.31.0

Flask-WTF==1.2.1
